[fix] 이름 수정
