+++ src/routes/orderRouter.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const { orderController } = require('../controllers');
+const { verificateToken } = require('../middlewares')
+const router = express.Router();
+
+router.get("/", verificateToken, orderController.getUserOrderController)
+
+
+module.exports = router;
