--- src/routes/orderlistRouter.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const orderlistController = require('../controllers/orderlistController');
-const router = express.Router();
-
-router.get("/check-order/:userId", orderlistController.userCheckOrderController)
-
-
-module.exports = router;
