+++ src/services/orderService.js
@@ -1,0 +1,9 @@
+const { orderDao } = require('../models');
+
+const getOrderByUserId = async (userId) => {
+  return await orderDao.getUserByOrder(userId);
+};
+
+module.exports = {
+  getOrderByUserId
+};
