--- src/services/orderlistService.js
@@ -1,9 +1,0 @@
-const { orderlistDao } = require('../models');
-
-const getCheckOrderList = async (userId) => {
-  return await orderlistDao.getUserCheckOrder(userId);
-};
-
-module.exports = {
-  getCheckOrderList
-};
